test(draftReview): consolidate draft review tab navigation tests and remove redundant file
